Se agrego el metodo de editar junto con su endpoint

+++ src/main/java/com/example/ThePetVerse/controller/PetController.java
@@ -45,14 +45,38 @@ public class PetController {
             }
             return ResponseEntity.badRequest().body(errors);
             }
+        Optional <Pet> petOp=this.petService.findByIDPet(pet.getId());
+        if(petOp.isPresent()){
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("The pet with the id" + pet.getId() + "already exists");
+        }
         Pet petSave=this.petService.savePet(pet);
         return ResponseEntity.status(HttpStatus.CREATED).body(petSave);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deletePet(@PathVariable Integer id){
+        Optional<Pet> petOp=this.petService.findByIDPet(id);
+        if(!petOp.isPresent()){
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("The pet with the id: "+id+" ISNT registred");
+        }
         this.petService.deletePet(id);
         return ResponseEntity.noContent().build();
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<?> editUser(@Validated @PathVariable Integer id, @RequestBody Pet petEdit, BindingResult result) {
+        if (result.hasErrors()) {
+            Map<String, String> errors = new HashMap<>();
+            for (FieldError error : result.getFieldErrors()) {
+                errors.put(error.getField(), error.getDefaultMessage());
+            }
+            return ResponseEntity.badRequest().body(errors);
+        }
+        Optional<Pet> userOp = this.petService.findByIDPet(id);
+        if (!userOp.isPresent()) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("The pet with the id " + id + " ISNT registred");
+        }
+        return ResponseEntity.ok(this.petService.editPet(id, petEdit));
     }
